before:
`rounded rect ...................... 0.515 ms      0.521 ms    ±0.007  x1000`
`svg decode ........................ 13.996 ms     14.122 ms    ±0.172   x354`


after:
`rounded rect ...................... 0.211 ms      0.215 ms    ±0.005  x1000`
`svg decode ........................ 13.873 ms     13.934 ms    ±0.049   x359`